Fix another set of conversion warnings

+++ Source/MediaStorageAndFileFormat/gdcmImageFragmentSplitter.cxx
@@ -76,7 +76,7 @@ bool ImageFragmentSplitter::Split()
   if( lastfrag )
     {
     Fragment splitfrag;
-    splitfrag.SetByteValue( p + nfrags * FragmentSizeMax, lastfrag );
+    splitfrag.SetByteValue( p + nfrags * FragmentSizeMax, (uint32_t)lastfrag );
     assert( nfrags * FragmentSizeMax + lastfrag == len );
     sq->AddFragment( splitfrag );
     }
