+++ Source/MediaStorageAndFileFormat/gdcmImageHelper.cxx
@@ -240,7 +240,7 @@ bool ComputeZSpacingFromIPP(const DataSet &ds, double &zspacing)
     meanspacing += current;
     prev = distances[i];
     }
-  meanspacing /= (nitems - 1);
+  meanspacing /= (double)(nitems - 1);
 
   //zspacing = distances[1] - distances[0];
   zspacing = meanspacing;
@@ -774,10 +774,10 @@ void ImageHelper::SetDimensionsValue(File& f, const Image & img)
   else
     {
     Attribute<0x0028,0x0010> rows;
-    rows.SetValue( dims[1] );
+    rows.SetValue( (uint16_t)dims[1] );
     ds.Replace( rows.GetAsDataElement() );
     Attribute<0x0028,0x0011> columns;
-    columns.SetValue( dims[0] );
+    columns.SetValue( (uint16_t)dims[0] );
     ds.Replace( columns.GetAsDataElement() );
     if( dims[2] > 1 )
       {
@@ -1087,7 +1087,7 @@ std::vector<double> ImageHelper::GetSpacingValue(File const & f)
         el.SetLength( entry.GetVM().GetLength() * entry.GetVR().GetSizeof() );
         el.Read( ss );
         assert( el.GetLength() == 2 );
-        for(unsigned long i = 0; i < el.GetLength(); ++i)
+        for(unsigned int i = 0; i < el.GetLength(); ++i)
           {
           if( el.GetValue(i) )
             {
@@ -1113,7 +1113,7 @@ std::vector<double> ImageHelper::GetSpacingValue(File const & f)
         ss.str( s );
         el.SetLength( entry.GetVM().GetLength() * entry.GetVR().GetSizeof() );
         el.Read( ss );
-        for(unsigned long i = 0; i < el.GetLength(); ++i)
+        for(unsigned int i = 0; i < el.GetLength(); ++i)
           sp.push_back( el.GetValue(i) );
         assert( sp.size() == (unsigned int)entry.GetVM() );
         }
@@ -1162,7 +1162,7 @@ std::vector<double> ImageHelper::GetSpacingValue(File const & f)
             ss.str( s );
             el.SetLength( entry.GetVM().GetLength() * entry.GetVR().GetSizeof() );
             el.Read( ss );
-            for(unsigned long i = 0; i < el.GetLength(); ++i)
+            for(unsigned int i = 0; i < el.GetLength(); ++i)
               {
               const double value = el.GetValue(i);
               sp.push_back( value );
@@ -1381,7 +1381,7 @@ void ImageHelper::SetSpacingValue(DataSet & ds, const std::vector<double> & spac
           assert( entry.GetVM() == VM::VM2 );
           for( unsigned int i = 0; i < entry.GetVM().GetLength(); ++i)
             {
-            el.SetValue( spacing[i], i );
+            el.SetValue( (int)spacing[i], i );
             }
           //assert( el.GetValue(0) == spacing[0] && el.GetValue(1) == spacing[1] );
           std::stringstream os;
@@ -1873,9 +1873,9 @@ bool ImageHelper::ComputeSpacingFromImagePositionPatient(const std::vector<doubl
     spacing[2] += z;
     }
   size_t n = imageposition.size() / 3;
-  spacing[0] /= n;
-  spacing[1] /= n;
-  spacing[2] /= n;
+  spacing[0] /= (double)n;
+  spacing[1] /= (double)n;
+  spacing[2] /= (double)n;
 
   return true;
 }
@@ -2049,7 +2049,7 @@ SmartPointer<LookupTable> ImageHelper::GetLUT(File const& f){
     // (0028,1101) US 0\0\16
     // (0028,1102) US 0\0\16
     // (0028,1103) US 0\0\16
-    const Tag tdescriptor(0x0028, (0x1101 + i));
+    const Tag tdescriptor(0x0028, (uint16_t)(0x1101 + i));
     //const Tag tdescriptor(0x0028, 0x3002);
     Element<VR::US,VM::VM3> el_us3 = {{ 0, 0, 0}};
     // Now pass the byte array to a DICOMizer:
@@ -2060,14 +2060,14 @@ SmartPointer<LookupTable> ImageHelper::GetLUT(File const& f){
     // (0028,1201) OW
     // (0028,1202) OW
     // (0028,1203) OW
-    const Tag tlut(0x0028, (0x1201 + i));
+    const Tag tlut(0x0028, (uint16_t)(0x1201 + i));
     //const Tag tlut(0x0028, 0x3006);
 
     // Segmented LUT
     // (0028,1221) OW
     // (0028,1222) OW
     // (0028,1223) OW
-    const Tag seglut(0x0028, (0x1221 + i));
+    const Tag seglut(0x0028, (uint16_t)(0x1221 + i));
     if( ds.FindDataElement( tlut ) )
       {
       const ByteValue *lut_raw = ds.GetDataElement( tlut ).GetByteValue();
